requirements/common.txt
Django==1.10
argparse==1.2.1
django-bootstrap3==7.0.1
django-ckeditor==5.0.3
gunicorn==19.6.0
wsgiref==0.1.2
